+++ ui-javafx/src/main/kotlin/edu/wpi/axon/ui/view/joblist/JobList.kt
@@ -1,4 +1,4 @@
-package edu.wpi.axon.ui.view
+package edu.wpi.axon.ui.view.joblist
 
 import de.jensd.fx.glyphs.fontawesome.FontAwesomeIcon
 import de.jensd.fx.glyphs.fontawesome.FontAwesomeIconView
@@ -14,35 +14,21 @@ import edu.wpi.axon.tfdata.optimizer.Optimizer
 import edu.wpi.axon.training.ModelDeploymentTarget
 import edu.wpi.axon.ui.ModelManager
 import edu.wpi.axon.ui.controller.JobBoard
-import edu.wpi.axon.ui.model.JobDto
 import edu.wpi.axon.ui.model.JobModel
 import javafx.collections.ListChangeListener
 import javafx.geometry.Orientation
-import javafx.scene.control.ProgressBar
 import javafx.scene.layout.Priority
-import org.koin.core.inject
-import tornadofx.ListCellFragment
 import tornadofx.ValidationMessage
 import tornadofx.ValidationSeverity
 import tornadofx.View
 import tornadofx.action
 import tornadofx.bindSelected
-import tornadofx.bindTo
-import tornadofx.booleanBinding
 import tornadofx.button
 import tornadofx.buttonbar
-import tornadofx.contextmenu
-import tornadofx.doubleBinding
 import tornadofx.enableWhen
 import tornadofx.field
 import tornadofx.insets
-import tornadofx.item
-import tornadofx.label
 import tornadofx.listview
-import tornadofx.onChange
-import tornadofx.progressbar
-import tornadofx.stringBinding
-import tornadofx.textarea
 import tornadofx.textfield
 import tornadofx.validator
 import tornadofx.vbox
@@ -157,64 +143,5 @@ class JobList : View() {
             target = ModelDeploymentTarget.Desktop,
             datasetPlugin = processMnistTypePlugin
         )
-    }
-}
-
-class JobListFragment : ListCellFragment<JobDto>() {
-    private val job = JobModel().bindTo(this)
-
-    override val root = vbox {
-        label(job.name.stringBinding(job.id) { "$it (${job.id.value})" })
-        progressbar(doubleBinding(job.status) {
-            value.let {
-                when (it) {
-                    TrainingScriptProgress.NotStarted -> 0.0
-                    TrainingScriptProgress.Creating -> ProgressBar.INDETERMINATE_PROGRESS
-                    TrainingScriptProgress.Initializing -> ProgressBar.INDETERMINATE_PROGRESS
-                    is TrainingScriptProgress.InProgress -> it.percentComplete
-                    TrainingScriptProgress.Completed -> 1.0
-                    is TrainingScriptProgress.Error -> 1.0
-                    else -> 0.0
-                }
-            }
-        }) {
-            maxWidth = Double.MAX_VALUE
-
-            job.status.onChange {
-                styleClass.setAll(
-                    when (it) {
-                        TrainingScriptProgress.NotStarted -> listOf("progress-bar")
-                        TrainingScriptProgress.Creating -> listOf("progress-bar")
-                        TrainingScriptProgress.Initializing -> listOf("progress-bar")
-                        is TrainingScriptProgress.InProgress -> listOf("progress-bar")
-                        TrainingScriptProgress.Completed -> listOf(
-                            "progress-bar-success",
-                            "progress-bar"
-                        )
-                        is TrainingScriptProgress.Error -> listOf(
-                            "progress-bar-error",
-                            "progress-bar"
-                        )
-                        else -> listOf("progress-bar-error", "progress-bar")
-                    }
-                )
-            }
-        }
-        contextmenu {
-            item("View Error") {
-                enableWhen(job.status.booleanBinding { it is TrainingScriptProgress.Error })
-                action {
-                    dialog(labelPosition = Orientation.VERTICAL) {
-                        field("Error Log") {
-                            textarea {
-                                text = (job.status.value as TrainingScriptProgress.Error).log
-                                isEditable = false
-                                isWrapText = true
-                            }
-                        }
-                    }
-                }
-            }
-        }
     }
 }
